requirements.txt
streamlit==1.40.0
langchain==0.3.7
langchain-community==0.3.7
langchain-openai==0.2.9
faiss-cpu==1.9.0
pypdf==5.1.0
python-dotenv==1.0.1
tiktoken==0.8.0
openai==1.54.4
httpx==0.27.2
plotly==6.5.2
pytesseract==0.3.10
pdf2image==1.17.0
